requirements.txt
transformers==4.52.4
datasets==3.6.0
bitsandbytes==0.46.0
trl==0.19.0
deepeval==3.1.8
peft==0.15.2
python-dotenv
sentencepiece